+++ projects/fhi-angular-highcharts/src/lib/services/diagram-type-group.service.ts
@@ -3,7 +3,7 @@ import { cloneDeep } from 'lodash-es';
 
 import { DiagramTypeIdValues } from '../constants-and-enums/diagram-type-ids';
 import { DiagramTypeGroup } from '../models/diagram-type-group.model';
-import { DiagramTypeGroups_NEW } from '../constants-and-enums/diagram-type-groups';
+import { DiagramTypeGroups } from '../constants-and-enums/diagram-type-groups';
 import { FhiDiagramSerie } from '../models/fhi-diagram-serie.model';
 import { FlaggedSerie } from '../models/flagged-serie.model';
 import { DiagramType } from '../models/diagram-type.model';
@@ -11,13 +11,24 @@ import { DiagramTypes } from '../constants-and-enums/fhi-diagram-types';
 
 @Injectable()
 export class DiagramTypeGroupService {
-  // TODO: The premise for the new diagramTypeGroups is that "map types" is implementend the same
-  //       way as "chart types", ie. that FhiDiagramOptions.mapTypeId is deprecated, and diferent
-  //       maps has it own type in DiagramTypes, not just one type with id "map" as is the case today.
   private activeDiagramType: DiagramType;
   private diagramTypeGroups!: DiagramTypeGroup[];
   private flaggedSeries!: FlaggedSerie[];
   private series!: FhiDiagramSerie[];
+
+  getActiveDiagramType(): DiagramType {
+    return this.activeDiagramType;
+  }
+
+  getActiveDiagramTypeGroup(): DiagramTypeGroup {
+    let activeGroup: DiagramTypeGroup;
+    this.diagramTypeGroups.forEach((group) => {
+      if (group.diagramType.active) {
+        activeGroup = group;
+      }
+    });
+    return activeGroup;
+  }
 
   getDiagramTypeGroups(): DiagramTypeGroup[] {
     return this.diagramTypeGroups;
@@ -25,7 +36,8 @@ export class DiagramTypeGroupService {
 
   updateDiagramTypeGroups(
     diagramTypeId: string,
-    diagramTypeSubset: string[] | undefined,
+    chartTypeSubset: string[] | undefined,
+    mapTypeSubset: string[] | undefined,
     flaggedSeries: FlaggedSerie[],
     series: FhiDiagramSerie[],
     diagramTypeGroups: DiagramTypeGroup[],
@@ -35,15 +47,15 @@ export class DiagramTypeGroupService {
     this.activeDiagramType = undefined;
     this.diagramTypeGroups = diagramTypeGroups
       ? cloneDeep(diagramTypeGroups)
-      : cloneDeep(DiagramTypeGroups_NEW);
+      : cloneDeep(DiagramTypeGroups);
 
-    this.loopGroupsAndUpdateDiagramTypes(diagramTypeSubset, diagramTypeId);
+    this.loopGroupsAndUpdateDiagramTypes(chartTypeSubset?.concat(mapTypeSubset), diagramTypeId);
     this.removeEmptyGroups();
     this.updateInactiveGroup();
     this.updateActiveGroup();
   }
 
-  diagramTypeDisabled(diagramTypeId: string): boolean {
+  diagramTypeIsDisabled(diagramTypeId: string): boolean {
     let disabled = false;
     this.diagramTypeGroups.forEach((group) => {
       group.children.forEach((diagramType) => {
@@ -139,7 +151,8 @@ export class DiagramTypeGroupService {
         diagramType.disabled = this.diagramTypeLineDisabled();
         break;
 
-      case DiagramTypeIdValues.map:
+      case DiagramTypeIdValues.mapFylker:
+      case DiagramTypeIdValues.mapFylker2019:
         diagramType.disabled = this.diagramTypeMapDisabled();
         break;