+++ projects/fhi-angular-highcharts/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Unreleased
 
-> May 22, 2024
+> May 27, 2024
 
+* :tada: **Enhancement** Update model `FhiDiagramOptions`: add more grouping of properties, rename some properties and remove some properties. The old properties still works, but will be deprecated in v5.
 * :bug: **Bugfix** Remove compact date formatting on chart diagrams
 
 ## 3.2.0